tests/wpt/web-platform-tests/websockets/Secure-Close-2999-reason.htm

<!DOCTYPE html>
<html>
<head>
    <title>W3C WebSocket API - Close Secure WebSocket - Code is 2999 and reason</title>
    <script type="text/javascript" src="/resources/testharness.js"></script>
    <script type="text/javascript" src="/resources/testharnessreport.js"></script>
    <script type="text/javascript" src="websocket.sub.js"></script>
</head>
<body>
    <div id="log"></div>
    <script type="text/javascript">

        var test = async_test("W3C WebSocket API - Create Secure WebSocket - Close the Connection - close(2999, reason) - INVALID_ACCESS_ERR is thrown");

        var wsocket = CreateWebSocket(true, false, false);

        wsocket.addEventListener('open', test.step_func(function (evt) {
            assert_throws("INVALID_ACCESS_ERR", function () { wsocket.close(2999, "Close not in range 3000-4999") });
            test.done();
        }), true);
    </script>

</body>
</html>
